tests/environment/requirements.txt
azure-core==1.26.1
azure-identity==1.12.0
azure-storage-blob==12.14.1
build==0.9.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.4
idna==3.4
importlib-metadata==5.1.0
isodate==0.6.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
packaging==22.0
pep517==0.13.0
portalocker==2.6.0
pycparser==2.21
PyJWT==2.6.0
pywin32==305
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
